Extra admin dashboard orders into separate component
